Refactor RPC servers and clients
