+++ src/rpc/farmer_rpc_client.py
@@ -1,13 +1,8 @@
-import aiohttp
-import asyncio
-
-from typing import Dict, Optional, List
-from src.util.byte_types import hexstr_to_bytes
-from src.types.sized_bytes import bytes32
-from src.util.ints import uint16
+from typing import Dict, List
+from src.rpc.rpc_client import RpcClient
 
 
-class FarmerRpcClient:
+class FarmerRpcClient(RpcClient):
     """
     Client to Chia RPC, connects to a local farmer. Uses HTTP/JSON, and converts back from
     JSON into native python objects before returning. All api calls use POST requests.
@@ -16,44 +11,5 @@ class FarmerRpcClient:
     to the full node.
     """
 
-    url: str
-    session: aiohttp.ClientSession
-    closing_task: Optional[asyncio.Task]
-
-    @classmethod
-    async def create(cls, port: uint16):
-        self = cls()
-        self.url = f"http://localhost:{str(port)}/"
-        self.session = aiohttp.ClientSession()
-        self.closing_task = None
-        return self
-
-    async def fetch(self, path, request_json):
-        async with self.session.post(self.url + path, json=request_json) as response:
-            response.raise_for_status()
-            return await response.json()
-
     async def get_latest_challenges(self) -> List[Dict]:
         return await self.fetch("get_latest_challenges", {})
-
-    async def get_connections(self) -> List[Dict]:
-        response = await self.fetch("get_connections", {})
-        for connection in response["connections"]:
-            connection["node_id"] = hexstr_to_bytes(connection["node_id"])
-        return response["connections"]
-
-    async def open_connection(self, host: str, port: int) -> Dict:
-        return await self.fetch("open_connection", {"host": host, "port": int(port)})
-
-    async def close_connection(self, node_id: bytes32) -> Dict:
-        return await self.fetch("close_connection", {"node_id": node_id.hex()})
-
-    async def stop_node(self) -> Dict:
-        return await self.fetch("stop_node", {})
-
-    def close(self):
-        self.closing_task = asyncio.create_task(self.session.close())
-
-    async def await_closed(self):
-        if self.closing_task is not None:
-            await self.closing_task
