+++ src/rpc/harvester_rpc_client.py
@@ -1,14 +1,9 @@
-import aiohttp
-import asyncio
-
 from blspy import PrivateKey, PublicKey
-from typing import Dict, Optional, List
-from src.util.byte_types import hexstr_to_bytes
-from src.types.sized_bytes import bytes32
-from src.util.ints import uint16
+from typing import Optional, List, Dict
+from src.rpc.rpc_client import RpcClient
 
 
-class HarvesterRpcClient:
+class HarvesterRpcClient(RpcClient):
     """
     Client to Chia RPC, connects to a local harvester. Uses HTTP/JSON, and converts back from
     JSON into native python objects before returning. All api calls use POST requests.
@@ -16,23 +11,6 @@ class HarvesterRpcClient:
     protocol on top of TCP), it's a separate protocol on top of HTTP thats provides easy access
     to the full node.
     """
-
-    url: str
-    session: aiohttp.ClientSession
-    closing_task: Optional[asyncio.Task]
-
-    @classmethod
-    async def create(cls, port: uint16):
-        self = cls()
-        self.url = f"http://localhost:{str(port)}/"
-        self.session = aiohttp.ClientSession()
-        self.closing_task = None
-        return self
-
-    async def fetch(self, path, request_json):
-        async with self.session.post(self.url + path, json=request_json) as response:
-            response.raise_for_status()
-            return await response.json()
 
     async def get_plots(self) -> List[Dict]:
         return await self.fetch("get_plots", {})
@@ -57,25 +35,3 @@ class HarvesterRpcClient:
             return await self.fetch(
                 "add_plot", {"filename": filename, "plot_sk": plot_sk_str}
             )
-
-    async def get_connections(self) -> List[Dict]:
-        response = await self.fetch("get_connections", {})
-        for connection in response["connections"]:
-            connection["node_id"] = hexstr_to_bytes(connection["node_id"])
-        return response["connections"]
-
-    async def open_connection(self, host: str, port: int) -> Dict:
-        return await self.fetch("open_connection", {"host": host, "port": int(port)})
-
-    async def close_connection(self, node_id: bytes32) -> Dict:
-        return await self.fetch("close_connection", {"node_id": node_id.hex()})
-
-    async def stop_node(self) -> Dict:
-        return await self.fetch("stop_node", {})
-
-    def close(self):
-        self.closing_task = asyncio.create_task(self.session.close())
-
-    async def await_closed(self):
-        if self.closing_task is not None:
-            await self.closing_task
